Refactor Python plugin
