+++ test/plugins/python.test.js
@@ -1,41 +1,41 @@
 import assert from 'assert';
-import pythonUniversal from '../../lib/resolver/python-universal.js';
+import python from '../../lib/plugins/python';
 import liveResolverQuery from '../../lib/resolver/live-resolver-query.js';
 
-describe('python-universal', () => {
+describe('python', () => {
   const path = '/octo/dog.py';
 
   it('resolves local file', () => {
     assert.deepEqual(
-      pythonUniversal({ path, target: '.foo' }),
+      python.resolve({ path, target: '.foo' }),
       '{BASE_URL}/octo/foo.py',
     );
   });
 
   it('resolves init file', () => {
     assert.deepEqual(
-      pythonUniversal({ path, target: '.' }),
+      python.resolve({ path, target: '.' }),
       '{BASE_URL}/octo/__init__.py',
     );
   });
 
   it('resolves package', () => {
     assert.deepEqual(
-      pythonUniversal({ path, target: 'foo' })[1],
+      python.resolve({ path, target: 'foo' })[1],
       liveResolverQuery({ type: 'pypi', target: 'foo' }),
     );
   });
 
   it('resolves scope package', () => {
     assert.deepEqual(
-      pythonUniversal({ path, target: 'foo.bar' })[1],
+      python.resolve({ path, target: 'foo.bar' })[1],
       liveResolverQuery({ type: 'pypi', target: 'foo' }),
     );
   });
 
   it('resolves buildin', () => {
     assert.deepEqual(
-      pythonUniversal({ path, target: 'foo' })[0],
+      python.resolve({ path, target: 'foo' })[0],
       'https://docs.python.org/3/library/foo.html',
     );
   });
